+++ Assets.xcassets/dashboard_FILL0_wght400_GRAD0_opsz48 1.imageset/Contents.json
@@ -1,0 +1,15 @@
+{
+  "images" : [
+    {
+      "filename" : "dashboard_FILL0_wght400_GRAD0_opsz48 1.svg",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  },
+  "properties" : {
+    "preserves-vector-representation" : true
+  }
+}
