+++ pureLocation/Controller/Mypage/DeleteTourViewController.swift
@@ -63,5 +63,12 @@ extension DeleteTourViewController : UICollectionViewDataSource {
     
     @objc func buttonTap (_ sender : UIButton) {
         sender.tag += 1
+        sender.tag += 1
+        if(sender.tag % 2 == 1) {
+            sender.setImage(UIImage(named: "check"), for: .normal)
+        }
+        else {
+            sender.setImage(UIImage(named: "DeleteImage"), for: .normal)
+        }
     }
 }
