+++ pureLocation/StoryBoards/MyPage.storyboard
@@ -92,13 +92,13 @@
                                                 <rect key="frame" x="0.0" y="0.0" width="393" height="97.666666666666671"/>
                                                 <subviews>
                                                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="돌굴러가유" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="li3-Qy-pnP">
-                                                        <rect key="frame" x="111" y="9.6666666666666714" width="86.666666666666686" height="24"/>
+                                                        <rect key="frame" x="111" y="28.666666666666671" width="86.666666666666686" height="24"/>
                                                         <fontDescription key="fontDescription" type="system" weight="semibold" pointSize="20"/>
                                                         <nil key="textColor"/>
                                                         <nil key="highlightedColor"/>
                                                     </label>
                                                     <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="circle.fill" catalog="system" translatesAutoresizingMaskIntoConstraints="NO" id="FAG-zm-5E7">
-                                                        <rect key="frame" x="24" y="2.6666666666666643" width="70" height="68.666666666666686"/>
+                                                        <rect key="frame" x="24" y="22.666666666666671" width="70" height="68.666666666666671"/>
                                                         <color key="tintColor" red="0.81960784310000001" green="0.81960784310000001" blue="0.83921568629999999" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                                                         <constraints>
                                                             <constraint firstAttribute="height" constant="70" id="1qp-mF-kJE"/>
@@ -106,13 +106,13 @@
                                                         </constraints>
                                                     </imageView>
                                                     <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="@gyulgyul" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="IBp-Ip-kKn">
-                                                        <rect key="frame" x="111" y="37.666666666666671" width="67.333333333333314" height="17"/>
+                                                        <rect key="frame" x="111" y="56.666666666666671" width="67.333333333333314" height="17"/>
                                                         <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                                         <color key="textColor" red="0.58039215690000001" green="0.58039215690000001" blue="0.58039215690000001" alpha="1" colorSpace="custom" customColorSpace="sRGB"/>
                                                         <nil key="highlightedColor"/>
                                                     </label>
                                                     <imageView clipsSubviews="YES" userInteractionEnabled="NO" contentMode="scaleAspectFit" horizontalHuggingPriority="251" verticalHuggingPriority="251" image="person" translatesAutoresizingMaskIntoConstraints="NO" id="cAW-dr-VbA">
-                                                        <rect key="frame" x="39" y="15.666666666666671" width="41" height="41"/>
+                                                        <rect key="frame" x="39" y="36.666666666666671" width="41" height="41"/>
                                                         <constraints>
                                                             <constraint firstAttribute="height" constant="41" id="ezv-OA-j1t"/>
                                                             <constraint firstAttribute="width" constant="41" id="rhb-2c-6mi"/>
@@ -122,13 +122,13 @@
                                                 <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                                                 <constraints>
                                                     <constraint firstItem="IBp-Ip-kKn" firstAttribute="top" secondItem="li3-Qy-pnP" secondAttribute="bottom" constant="4" id="51j-RO-fQ6"/>
-                                                    <constraint firstAttribute="bottom" secondItem="cAW-dr-VbA" secondAttribute="bottom" constant="41" id="BwA-0j-Jqj"/>
-                                                    <constraint firstAttribute="bottom" secondItem="FAG-zm-5E7" secondAttribute="bottom" constant="26" id="D7w-0i-FO7"/>
+                                                    <constraint firstAttribute="bottom" secondItem="cAW-dr-VbA" secondAttribute="bottom" constant="20" id="BwA-0j-Jqj"/>
+                                                    <constraint firstAttribute="bottom" secondItem="FAG-zm-5E7" secondAttribute="bottom" constant="6" id="D7w-0i-FO7"/>
                                                     <constraint firstItem="cAW-dr-VbA" firstAttribute="leading" secondItem="ejx-gU-2eK" secondAttribute="leading" constant="39" id="Sef-zy-HBF"/>
                                                     <constraint firstItem="li3-Qy-pnP" firstAttribute="leading" secondItem="FAG-zm-5E7" secondAttribute="trailing" constant="17" id="Ttg-25-Mvp"/>
                                                     <constraint firstItem="IBp-Ip-kKn" firstAttribute="leading" secondItem="FAG-zm-5E7" secondAttribute="trailing" constant="17" id="qeR-2l-914"/>
                                                     <constraint firstItem="FAG-zm-5E7" firstAttribute="leading" secondItem="ejx-gU-2eK" secondAttribute="leading" constant="24" id="wAF-aW-fnX"/>
-                                                    <constraint firstAttribute="bottom" secondItem="li3-Qy-pnP" secondAttribute="bottom" constant="64" id="zxo-yb-jep"/>
+                                                    <constraint firstAttribute="bottom" secondItem="li3-Qy-pnP" secondAttribute="bottom" constant="45" id="zxo-yb-jep"/>
                                                 </constraints>
                                             </view>
                                         </subviews>
@@ -776,7 +776,7 @@
         <image name="circle.fill" catalog="system" width="128" height="123"/>
         <image name="hart" width="16" height="16"/>
         <image name="insta" width="28" height="28"/>
-        <image name="map" width="28" height="28"/>
+        <image name="map" width="25" height="24"/>
         <image name="person" width="41" height="41"/>
         <image name="profile2" width="50" height="50"/>
         <systemColor name="systemBackgroundColor">
